requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.4
